Migra add_xp_group() para módulo char e drop_random_prize() para item
